% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_table.R
\name{descriptive_table}
\alias{descriptive_table}
\title{Generate html table of descritive statistics for data}
\usage{
descriptive_table(dataframe, num, char, grp, title = "Table",
  label.option = F, option = F, digit = 1, num.test = c("None",
  "Wilcoxon/Kruskal", "ANOVA"))
}
\arguments{
\item{dataframe}{Data to be used}

\item{num}{A vector of numeric variable names}

\item{char}{A vector of categorical variable names}

\item{grp}{Group names, limited to one name}

\item{title}{Title of the table}

\item{label.option}{Default is \code{'FALSE'}.\code{'TURE'} uses variable label rather than variable name}

\item{option}{Default is \code{'FALSE'}. \code{'TRUE'} uses min and max for numeric variable and percentage for categorical variable}

\item{digit}{Round the numeric statistics except for p-value. P-value always had 3 decimals.}

\item{num.test}{A character string specifying the test method for numeric variable among groups, 
must be one of "None", "ANOVA", "Wilcoxon/Kruskal". "Wilcoxon/Kruskal" means if the group variable has two groups, 
then use Wilcoxon sum rank test, if has more than two groups, use Kruskal-Wallis test.}
}
\value{
Desciptive tables for selected variables, median (min-max) for numeric variable, n(\%) for characteristic
 variables. ANOVA for numeric variables with normality assumption; Wilcoxon test for non-parametric variable and 
 two groups; Kruiskal test for more than two groups (excluding missing group level) Chi-square test for 
 categorical variables (excluding missing levels, both variable and group)
}
\description{
Generate html table of descritive statistics for data
}
