% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcut.R
\name{findcut}
\alias{findcut}
\title{Find optimal cutpoints for continuous covariates based on the method
proposed by Contal and O'Quigley}
\usage{
findcut(dataframe, survtime, survevent, cutvar)
}
\arguments{
\item{dataframe}{Data to be used}

\item{survtime}{Survival time}

\item{survevent}{Survival event}

\item{cutvar}{Variable to be dichotomized}
}
\description{
Find optimal cutpoints for continuous covariates based on the method
proposed by Contal and O'Quigley
}
\references{
Contal and O’Quigley, 1999
Williams et al., 2006
}
